Split Main.cs into separate classes
